+++ app/Http/Controllers/Api/subcategoriesController.php
@@ -99,11 +99,14 @@ class subcategoriesController extends Controller
             return response()->json($this->messages['not_found'], 404);
         }
 
-        // Eliminar la subcategoría encontrada
-        $subcategory->delete();
-
-        // Retornar respuesta JSON con mensaje de éxito y código de estado 200
-        return response()->json($this->messages['deleted'], 200);
+         // Verificar si se encuentra el producto antes de eliminar
+        if ($subcategory) {
+            // Eliminar el producto encontrado
+            $subcategory->delete();
+            
+            // Retornar respuesta JSON con mensaje de éxito y código de estado 200
+            return response()->json($this->messages['deleted'], 200);
+        }
     }
